# src/main/resources/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: "Bitemporal Person-Address-Agreement Service"
  description: |
    REST API for managing persons, addresses, and agreements with full
    bitemporal (valid-time + transaction-time) support as described by
    Richard Snodgrass.

    ## Time Travel
    Use `validAt` and `transactionAt` query parameters on any GET endpoint
    to query the state of an entity at a specific point in the bitemporal
    plane.

    ## Bitemporal Model
    - **Valid Time**: When the fact was true in the real world (`validFrom` / `validTo`)
    - **Transaction Time**: When the fact was recorded in the database (`transactionFrom` / `transactionTo`)

    ## API Key Permissions
    All requests must include the `X-API-Key` header:
    - **READ_WRITE**: full access (GET, POST, PUT, DELETE)
    - **READ**: read-only access (GET only)
    - **NO_ACCESS**: all requests denied
  version: "1.0.0"
  contact:
    name: "BSoft Development Team"
servers:
  - url: "http://localhost:8080/api/v1"
    description: "Local development server"
  - url: "https://api.example.com/api/v1"
    description: "Production server"
security:
  - ApiKeyAuth: [ ]
...
paths:
...
  /persons:
    get:
      tags: [ persons ]
      operationId: listPersons
      summary: List persons at a given bitemporal point in time
      parameters:
        - $ref: '#/components/parameters/ValidAt'
        - $ref: '#/components/parameters/TransactionAt'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Paginated list of persons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [ persons ]
      operationId: createPerson
      summary: Create a new person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonRequest'
      responses:
        '201':
          description: Person created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /persons/{personId}:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [ persons ]
      operationId: getPerson
      summary: Get a person at a given bitemporal point in time
      parameters:
        - $ref: '#/components/parameters/ValidAt'
        - $ref: '#/components/parameters/TransactionAt'
      responses:
        '200':
          description: Person found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [ persons ]
      operationId: updatePerson
      summary: Update a person (creates a new transaction-time version)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonRequest'
      responses:
        '200':
          description: Person updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [ persons ]
      operationId: terminatePerson
      summary: Terminate the valid period of a person
      description: Ends the person's valid period. The record is preserved in transaction time for auditing.
      parameters:
        - name: validTo
          in: query
          required: false
          description: End of valid period (defaults to now)
          schema:
            type: string
            format: date-time
      responses:
        '204':
          description: Person valid period terminated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /persons/{personId}/history:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [ persons ]
      operationId: getPersonHistory
      summary: Get the complete bitemporal history of a person
      description: Returns all versions across both valid time and transaction time.
      responses:
        '200':
          description: Full bitemporal history ordered by transactionFrom, validFrom
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /persons/{personId}/addresses:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [ person-relations ]
      operationId: getPersonAddresses
      summary: Get addresses linked to a person at a given bitemporal point
      parameters:
        - $ref: '#/components/parameters/ValidAt'
        - $ref: '#/components/parameters/TransactionAt'
      responses:
        '200':
          description: List of person-address relations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonAddress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [ person-relations ]
      operationId: addPersonAddress
      summary: Link an address to a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonAddressRequest'
      responses:
        '201':
          description: Relation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonAddress'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /persons/{personId}/addresses/{addressId}:
    parameters:
      - $ref: '#/components/parameters/PersonId'
      - $ref: '#/components/parameters/AddressId'
    delete:
      tags: [ person-relations ]
      operationId: removePersonAddress
      summary: Terminate a person-address relation
      parameters:
        - name: validTo
          in: query
          required: false
          description: End of valid period (defaults to now)
          schema:
            type: string
            format: date-time
      responses:
        '204':
          description: Relation terminated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /persons/{personId}/agreements:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [ person-relations ]
      operationId: getPersonAgreements
      summary: Get agreements linked to a person at a given bitemporal point
      parameters:
        - $ref: '#/components/parameters/ValidAt'
        - $ref: '#/components/parameters/TransactionAt'
      responses:
        '200':
          description: List of person-agreement relations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonAgreement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [ person-relations ]
      operationId: addPersonAgreement
      summary: Link an agreement to a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonAgreementRequest'
      responses:
        '201':
          description: Relation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonAgreement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /persons/{personId}/agreements/{agreementId}:
    parameters:
      - $ref: '#/components/parameters/PersonId'
      - $ref: '#/components/parameters/AgreementId'
    delete:
      tags: [ person-relations ]
      operationId: removePersonAgreement
      summary: Terminate a person-agreement relation
      parameters:
        - name: validTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '204':
          description: Relation terminated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== ADDRESSES ====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Three permission levels:
        - READ_WRITE: full CRUD access
        - READ: read-only (GET) access
        - NO_ACCESS: all requests denied (key is revoked)
  parameters:
    PersonId:
      name: personId
      in: path
      required: true
      description: Person business UUID (stable across versions)
      schema:
        type: string
        format: uuid
    AddressId:
      name: addressId
      in: path
      required: true
      description: Address business UUID (stable across versions)
      schema:
        type: string
        format: uuid
    AgreementId:
      name: agreementId
      in: path
      required: true
      description: Agreement business UUID (stable across versions)
      schema:
        type: string
        format: uuid
    ValidAt:
      name: validAt
      in: query
      required: false
      description: |
        Point-in-time for the valid time dimension (ISO 8601 date-time).
        Defaults to current time — returns the currently valid state.
      schema:
        type: string
        format: date-time
    TransactionAt:
      name: transactionAt
      in: query
      required: false
      description: |
        Point-in-time for the transaction time dimension (ISO 8601 date-time).
        Defaults to current time — returns the state as currently recorded.
      schema:
        type: string
        format: date-time
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    BitemporalInfo:
      type: object
      description: Bitemporal metadata attached to every versioned record
      properties:
        validFrom:
          type: string
          format: date-time
          description: Start of valid period (when fact became true in reality)
        validTo:
          type: string
          format: date-time
          description: End of valid period (null = currently valid)
        transactionFrom:
          type: string
          format: date-time
          description: When this version was recorded in the database
        transactionTo:
          type: string
          format: date-time
          description: When this version was superseded (null = current version)

    # ==================== PERSON ====================
    Person:
      allOf:
        - $ref: '#/components/schemas/BitemporalInfo'
        - type: object
          required: [ uid, firstName, lastName, dateOfBirth ]
          properties:
            uid:
              type: string
              format: uuid
              description: Stable business identifier (same UUID across all versions)
            firstName:
              type: string
              maxLength: 100
            lastName:
              type: string
              maxLength: 100
            dateOfBirth:
              type: string
              format: date
    PersonRequest:
      type: object
      required: [ firstName, lastName, dateOfBirth ]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        dateOfBirth:
          type: string
          format: date
        validFrom:
          type: string
          format: date-time
          description: Start of valid period (defaults to now)
        validTo:
          type: string
          format: date-time
          description: End of valid period (defaults to open/infinity)
    PersonPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Person'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer

    # ==================== ADDRESS ====================
    Address:
      allOf:
        - $ref: '#/components/schemas/BitemporalInfo'
        - type: object
          required: [ uid, street, houseNumber, postalCode, city ]
          properties:
            uid:
              type: string
              format: uuid
            street:
              type: string
              maxLength: 200
            houseNumber:
              type: string
              maxLength: 20
            postalCode:
              type: string
              maxLength: 20
            city:
              type: string
              maxLength: 100
...
    AgreementState:
      type: string
      enum: [ NEW, CHANGED, READY ]
      description: Lifecycle state of an agreement
    Agreement:
      allOf:
        - $ref: '#/components/schemas/BitemporalInfo'
        - type: object
          required: [ uid, title, state ]
          properties:
            uid:
              type: string
              format: uuid
            title:
              type: string
              maxLength: 200
            description:
              type: string
            state:
              $ref: '#/components/schemas/AgreementState'
...
    AddressType:
      type: string
      enum: [ HOME, WORK ]
      description: The role of the address for the person
    PersonAddress:
      type: object
      properties:
        uid:
          type: string
          format: uuid
        personUid:
          type: string
          format: uuid
        addressUid:
          type: string
          format: uuid
        addressType:
          $ref: '#/components/schemas/AddressType'
        address:
          $ref: '#/components/schemas/Address'
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        transactionFrom:
          type: string
          format: date-time
        transactionTo:
          type: string
          format: date-time
    PersonAddressRequest:
      type: object
      required: [ addressUid, addressType ]
      properties:
        addressUid:
          type: string
          format: uuid
        addressType:
          $ref: '#/components/schemas/AddressType'
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
    PersonAgreement:
      type: object
      properties:
        uid:
          type: string
          format: uuid
        personUid:
          type: string
          format: uuid
        agreementUid:
          type: string
          format: uuid
        agreement:
          $ref: '#/components/schemas/Agreement'
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        transactionFrom:
          type: string
          format: date-time
        transactionTo:
          type: string
          format: date-time
    PersonAgreementRequest:
      type: object
      required: [ agreementUid ]
      properties:
        agreementUid:
          type: string
          format: uuid
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time

    # ==================== ERRORS ====================
    ApiError:
      type: object
      required: [ status, message, timestamp ]
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Bad request — validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Forbidden — insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found at the requested bitemporal point
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Conflict — resource already exists or constraint violated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'